requirements.txt
langchain
langchain-community
langchain-groq
langchain-text-splitters
pypdf
docx2txt
fastapi[all]
python-dotenv
pymongo[srv]
mongoengine
passlib[bcrypt]>=1.7.4
email-validator
python-jose[cryptography]
fastapi-mail
beautifulsoup4
slowapi>=0.1.9
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

